app/core/config.py
import os
from dotenv import load_dotenv
import cloudinary
load_dotenv()

JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
JWT_ALGORITHM = os.getenv('JWT_ALGORITHM')
MONGO_URL = os.getenv('MONGO_URL')
ORIGIN = os.getenv('ORIGIN')

cloudinary.config(
    cloud_name=os.getenv("CLOUDINARY_NAME"),
    api_key=os.getenv("CLOUDINARY_KEY"),
    api_secret=os.getenv("CLOUDINARY_SECRET")
)
